requirements.txt
aiogram==3.15.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
asyncpg==0.30.0
aiosqlite==0.20.0
greenlet==3.1.1